# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/containers/{visitId}/holds": {
      "get": {
        "operationId": "OperationalHoldsController_findManyForVisit",
        "parameters": [
          {
            "name": "visitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OperationalHolds"
        ]
      },
      "post": {
        "operationId": "OperationalHoldsController_create",
        "parameters": [
          {
            "name": "visitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOperationalHoldDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "OperationalHolds"
        ]
      }
    },
    "/api/operational-holds/{holdId}": {
      "get": {
        "operationId": "OperationalHoldsController_findById",
        "parameters": [
          {
            "name": "holdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OperationalHolds"
        ]
      }
    },
    "/api/containers/{visitId}/holds/{holdId}/release": {
      "post": {
        "operationId": "OperationalHoldsController_release",
        "parameters": [
          {
            "name": "visitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "holdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseOperationalHoldDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "OperationalHolds"
        ]
      }
    },
...
  },
  "info": {
    "title": "ICD Management API",
    "description": "Production-grade ICD Container & Yard Management Backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT Access Token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateOperationalHoldDto": {
        "type": "object",
        "properties": {}
      },
      "ReleaseOperationalHoldDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}